requirements.txt
pyrogram>=2.0.0
tgcrypto
ollama
python-dotenv
rich
